requirements.txt
PyYAML>=5.4
openai>=0.27.0
